Scrape/House_Scraper/requirements.txt
psycopg2==2.8.6
openpyxl==3.0.7
requests==2.25.1
python-telegram-bot==13.5